Exclusão de contatos implementada e adequação em funções anteriores

+++ src/main/java/com/gestaodecontatos/packer/Empacotador.java
@@ -19,16 +19,19 @@ public class Empacotador {
     }
 
     public static void serializar(File arq, ContatoCollection contatos) throws Exception {
-        arq.delete();
-        arq.createNewFile();
-        ObjectOutputStream obj = new ObjectOutputStream(new FileOutputStream(arq));
-        obj.writeObject(contatos.getContatos());
-        obj.close();
-
+        if (arq.exists()) { // Se o arquivo existe, deleta o mesmo para evitar sobreposição dos dados
+            arq.delete();
+        }
+        if (!contatos.getContatos().isEmpty()) { //Se o ArrayList contendo os contatos tiver vazio, salva os contatos em arquivo, caso contrário, apenas deleta o arquivo
+            arq.createNewFile();
+            ObjectOutputStream obj = new ObjectOutputStream(new FileOutputStream(arq));
+            obj.writeObject(contatos.getContatos());
+            obj.close();
+        }
     }
 
     public static ArrayList<Contato> deserializar(File arq) throws Exception {
-        ArrayList<Contato> contatos = new ArrayList<Contato>();
+        ArrayList<Contato> contatos = new ArrayList<>();
 
         ObjectInputStream obj = new ObjectInputStream(new FileInputStream(arq));
         contatos = (ArrayList<Contato>) obj.readObject();
